refactor: ignore entire folder instead of each files
style: update comment
