+++ src/components/layouts/ReleaseLayout/index.jsx
@@ -1,5 +1,5 @@
-import { ReleasePage } from '@site/src/pages/releases/_index';
-import { ViewRelease } from '@site/src/pages/releases/_view';
+import { ReleasePage } from '@site/src/pages/_releases/index';
+import { ViewRelease } from '@site/src/pages/_releases/view';
 import Layout from '@theme/Layout';
 import NotFound from '@theme/NotFound';
 import React from 'react';
