+++ src/pages/_releases/view.jsx
